requirements.txt
sh
Jinja2
colorama #https://github.com/tartley/colorama
futures
python-dotenv
